refactor(session): extract handler modules and pipeline utilities from LlmSessionActor (#411)
Extract 9 focused modules from LlmSessionActor, reducing it from 3,208 to
2,273 lines (~930 lines relocated).

Handler modules (src/Netclaw.Actors/Sessions/Handlers/):
- TurnStateTracker: per-turn counters, duplicate detection, budget tracking
- ProcessingWatchdog: operation ID tracking, timer management, expiry validation
- DeliveryRetryHandler: retry counting, eligibility, nudge construction
- DiscoveredToolCache: MCP tool retention, lease countdown, eviction

Pipeline utilities (src/Netclaw.Actors/Sessions/Pipelines/):
- SessionTitleGenerator: title generation decision + sidecar LLM call
- SessionCompactionPipeline: tiered compaction, observation generation,
  token estimation, CompactionParameters record
- SessionLlmInvoker: streaming LLM invocation with delta forwarding
- SessionToolExecutionPipeline: parallel tool execution, result clamping,
  sub-agent finding review, ToolCallResult record
- SessionRecallManager: turn recall cache, progressive exclusion, recall
  injection and formatting

Each module is independently testable without an ActorSystem. The actor
delegates to modules via field references, preserving identical behavior.

All 1,400 tests pass.
