+++ src/Netclaw.Actors/Sessions/Pipelines/SessionLlmInvoker.cs
@@ -1,0 +1,140 @@
+using System.Text;
+using Akka.Actor;
+using Microsoft.Extensions.AI;
+using AiChatMessage = Microsoft.Extensions.AI.ChatMessage;
+
+namespace Netclaw.Actors.Sessions.Pipelines;
+
+/// <summary>
+/// Async pipeline for invoking the LLM. Runs on the thread pool and sends
+/// results back to the session actor via <c>self.Tell()</c>.
+/// </summary>
+internal static class SessionLlmInvoker
+{
+    public static async Task InvokeAsync(
+        IChatClient client,
+        List<AiChatMessage> messages,
+        ChatOptions? options,
+        IActorRef self,
+        TimeSpan timeout)
+    {
+        try
+        {
+            using var cts = new CancellationTokenSource(timeout);
+            var response = await StreamAsync(client, messages, options, self, cts.Token);
+            self.Tell(response);
+        }
+        catch (OperationCanceledException ex)
+        {
+            self.Tell(new LlmCallFailed
+            {
+                Cause = new TimeoutException(
+                    $"LLM call exceeded timeout of {timeout.TotalSeconds:F0}s",
+                    ex)
+            });
+        }
+        catch (Exception ex)
+        {
+            self.Tell(new LlmCallFailed { Cause = ex });
+        }
+    }
+
+    public static async Task<LlmResponseReceived> StreamAsync(
+        IChatClient client,
+        List<AiChatMessage> messages,
+        ChatOptions? options,
+        IActorRef self,
+        CancellationToken cancellationToken)
+    {
+        var contents = new List<AIContent>();
+        var updates = new List<ChatResponseUpdate>();
+        var textBuilder = new StringBuilder();
+        var thinkingBuilder = new StringBuilder();
+        string? pendingTextDelta = null;
+        string? pendingThinkingDelta = null;
+        var textDeltaCount = 0;
+        var thinkingDeltaCount = 0;
+
+        await foreach (var update in client.GetStreamingResponseAsync(messages, options, cancellationToken))
+        {
+            updates.Add(update);
+
+            if (update.Contents is not null)
+            {
+                foreach (var content in update.Contents)
+                {
+                    switch (content)
+                    {
+                        case TextContent text when !string.IsNullOrEmpty(text.Text):
+                            textBuilder.Append(text.Text);
+                            textDeltaCount++;
+                            if (textDeltaCount == 1)
+                            {
+                                pendingTextDelta = text.Text;
+                            }
+                            else
+                            {
+                                if (textDeltaCount == 2 && !string.IsNullOrEmpty(pendingTextDelta))
+                                {
+                                    self.Tell(new LlmResponseDeltaReceived
+                                    {
+                                        Content = new TextContent(pendingTextDelta)
+                                    });
+                                }
+
+                                self.Tell(new LlmResponseDeltaReceived { Content = content });
+                            }
+                            break;
+
+                        case TextReasoningContent thinking when !string.IsNullOrEmpty(thinking.Text):
+                            thinkingBuilder.Append(thinking.Text);
+                            thinkingDeltaCount++;
+                            if (thinkingDeltaCount == 1)
+                            {
+                                pendingThinkingDelta = thinking.Text;
+                            }
+                            else
+                            {
+                                if (thinkingDeltaCount == 2 && !string.IsNullOrEmpty(pendingThinkingDelta))
+                                {
+                                    self.Tell(new LlmResponseDeltaReceived
+                                    {
+                                        Content = new TextReasoningContent(pendingThinkingDelta)
+                                    });
+                                }
+
+                                self.Tell(new LlmResponseDeltaReceived { Content = content });
+                            }
+                            break;
+
+                        case FunctionCallContent:
+                            contents.Add(content);
+                            break;
+                    }
+                }
+            }
+
+        }
+
+        if (thinkingBuilder.Length > 0)
+            contents.Add(new TextReasoningContent(thinkingBuilder.ToString()));
+
+        if (textBuilder.Length > 0)
+            contents.Add(new TextContent(textBuilder.ToString()));
+
+        var response = updates.Count > 0
+            ? updates.ToChatResponse()
+            : new ChatResponse(new AiChatMessage(ChatRole.Assistant, contents));
+
+        if (response.Messages.Count == 0)
+            response.Messages.Add(new AiChatMessage(ChatRole.Assistant, contents));
+
+        return new LlmResponseReceived
+        {
+            Response = response,
+            StreamedText = textDeltaCount > 1,
+            StreamedThinking = thinkingDeltaCount > 1,
+            RecallResult = null
+        };
+    }
+}
